Rename test cases to be more descriptive
