+++ src/adapter/etl-adapter-elasticsearch/tests/Flow/ETL/Adapter/Elasticsearch/Tests/Integration/ElasticsearchIntegrationTest.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Flow\ETL\Adapter\Elasticsearch\Tests\Integration\ElasticsearchPHP;
+namespace Flow\ETL\Adapter\Elasticsearch\Tests\Integration;
 
+use Flow\ETL\{Flow, Row, Rows};
+use Flow\ETL\Adapter\Elasticsearch\EntryIdFactory\EntryIdFactory;
 use function Flow\ETL\Adapter\Elasticsearch\{es_hits_to_rows, from_es, to_es_bulk_index};
 use function Flow\ETL\DSL\{bool_entry, int_entry, string_entry};
-use Flow\ETL\Adapter\Elasticsearch\EntryIdFactory\EntryIdFactory;
-use Flow\ETL\{Adapter\Elasticsearch\Tests\Integration\TestCase, Flow, Row, Rows};
 
-final class IntegrationTest extends TestCase
+final class ElasticsearchIntegrationTest extends ElasticsearchTestCase
 {
     public const DESTINATION_INDEX = 'etl-test-destination-index';
