+++ src/adapter/etl-adapter-elasticsearch/tests/Flow/ETL/Adapter/Elasticsearch/Tests/Integration/ElasticsearchPHP/ElasticsearchLoaderTest.php
@@ -6,9 +6,9 @@ namespace Flow\ETL\Adapter\Elasticsearch\Tests\Integration\ElasticsearchPHP;
 
 use function Flow\ETL\Adapter\Elasticsearch\{entry_id_factory, hash_id_factory, to_es_bulk_index, to_es_bulk_update};
 use function Flow\ETL\DSL\{generate_random_string, string_entry};
-use Flow\ETL\{Adapter\Elasticsearch\Tests\Integration\TestCase, Config, FlowContext, Row, Rows};
+use Flow\ETL\{Adapter\Elasticsearch\Tests\Integration\ElasticsearchTestCase, Config, FlowContext, Row, Rows};
 
-final class ElasticsearchLoaderTest extends TestCase
+final class ElasticsearchLoaderTest extends ElasticsearchTestCase
 {
     public const INDEX_NAME = 'etl-test-index';
